Rewrite how handling IQ requests works
We are now using coroutines, one per IQ request. This allows for
parallel processing of IQs (which is generally desirable) and more
flexiblity.

Coroutines are the most abstract object we currently have: It can be a
normal function or be a real coroutine. Thus, any primitive can be used
with a coroutine, with the usual asyncio contract (don’t block for too
long).