Implement make_reply in xml instead of stanza
This change obsoletes the ugly hack we had before, where we would
temporarily add the reply node to the original node to make it inherit
the correct parser. This is wrong for several reasons.

This change also requires the constructor of a reply to know a TX
context, but that should always be available.