Implement a response action which raises BangSuccess
Can be used to prematurely abort a test. Useful to test if the client
e.g. chooses the correct SASL mechanism, and the SASL exchange itself is
known to work. There is no use to duplicate the SASL exchange here, just
put a "!success" instead of the response node sequence.