Fix stream management state desync when receiving errornous stanzas
The first and only test in tests/test_highlevel.py tests against this.
Before this patch, the counter would not increase for the errornous
stanza as it was handled by the XMLStream protocol itself; that way, the
StanzaStream could not know that a stanza was received and that the
counter needed to be incremented.

The XMLStream now forwards the errornous stanza alongside with an
exception to the StanzaStream, which can act accordingly.