Implement byte-level debug output in XMLStream
Previously we just serialised the XSOs in a separate buffer; this is
inaccurate with respect to the use of namespace prefixes, as the objects
are serialised in different context for debugging than for actually
sending them.

The new debug facilities hook into the write method passed to the actual
XSO serialiser, thus providing accurate output.