stanza: fix class of exception thrown by IQ.validate
ValueError is *not* handled by the XMLStream and will kill it.
This is intentional. Only specific errors are handled and treated
as recoverable parser errors. Things like ValueError of unknown
origin most likely indicate a bug in aioxmpp.

Thus, we wrap validation errors in stanza.StanzaError to make
the XMLStream treat it like any other stanza deserialisation
error (by invoking the streams error handler and continuing
with business as usual).