node: introduce aioxmpp.Client.on_stream_resumed
Allows a user/application/service to see when the stream is usable
again after on_stream_suspended. This is useful for things which
are not sensible to do while the stream is suspended, for example,
sending pings.

I don’t know why we didn’t have this in the past.