muc: self_ping: enforce that timedelta properties are set with correct type
The errors which occur when the timeout properties are assigned with
a value which is not compatible with timedelta() are non-obvious and
hard to catch.

Thus, we enforce compatibility with timedelta() at the time of
assignment, by attempting to add the neutral element (w.r.t. addition)
(i.e. timedelta() i.e. zero) to the value.