xml: fix serialisation of re-ocurring nested namespaces
This fixes an oversight where namespace URIs whose prefix has
been re-bound would not be removed from _curr_ns_map. That leads
to the prefix/namespace not being re-declared if the namespace was
declared further up in the hierarchy, but the prefix used in the
initial declaration was re-bound in-between.

Please see the added test cases for example situations.

Fixes #295.