docs: xso: Re-write much of the XSO documentation
This focuses on integrating nicely with the new user guide chapter
(hence the re-writes of the first lines of all the XSO descriptor
and type classes). In addition, I upgraded some of the docstrings
to the newer format.

There’s now also a quick intro section in the XSO subpackage, which
should be non-disruptive to power users (who’ll have to use the
ToC anyways) and very useful for newbies (who’ll be linked to the
user guide, hopefully, as a nice intro).