xso: fix assignment of enum members when only allow_coerce is true
Previously, the code would force coercion of the value through
the *nested* type (in most cases, String) first even if the value
was already an enumeration member. This obviously breaks because
the enumeration member is an instance of the Enumeration and not of
any other type.

This bug occurs only if deprecate_coerce is *false*, because
otherwise there was an additional check for enum membership in the
code path.