muc: treat remote server unreachability like timeout self-ping
If the remote server is unreachable network-wise, we cannot know
whether we have been removed. In addition, a rejoin will fail
deterministically.

It is thus better to stay in the undecided pinging state (treat it
like a timeout): this will allow us to wait for a rejoin until we
know more about the current state and the server is actually
reachable again.

See-Also: https://github.com/xsf/xeps/pull/834
See-Also: #312