protocol: add support for multiple and late-created features_future:s
The idea behind this is that this allows us to create an XMLStream
early and configure it (e.g. with future language settings, or
namespaces or whatever) before passing it down to the place where
the actual stream negotiation happens.

While we could instead pass down the future, that would be more
error-prone (and less flexible) than passing down the XMLStream
itself.