misc: fix stanza ID declaration to not break everything
This was helpfully found by the OpenFire CI [1], where they actually
test against aioxmpp devel, so kudos to them. I should've waited with
the release for a day until their jobs kicked in ;).

Either way, the problem is that MultiDict only accepts strings as keys
-- not JIDs, as we tried to put into them.

As there's no sane way out of that (stringifying the JIDs would be
terrible), we now use a ChildList instead. This is a breaking API
change, but we just yanked 0.13.0 off PyPI and the Debian package wasn't
ready anyway, so we'll put this in a 0.13.1.

   [1]: https://github.com/igniterealtime/Openfire/